% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{clusterMethods}
\alias{clusterMethods}
\alias{clusters}
\alias{mat}
\alias{nClusters}
\alias{summary,clena-method}
\title{Basic methods for a clena object.}
\usage{
clusterMethods(object)

nClusters(object)

clusters(object, method)

mat(object)

\S4method{summary}{clena}(object, digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{object}{a clena object}

\item{method}{as clMethods in clena function}
}
\value{
clusterMethods: a character vector.

nClusters: a numeric vector.

clusters: a list or hclust depends on the method

mat: a matrix

summary: a summary of a clena object.
}
\description{
clusterMethods: get the methods of clustering used.

nClusters: get the number of clusters from a clena object.

clusters: get the cluster of a certain clustering method.

mat: get the original data from a clena object.

summary: a summary of a clena object.
}
\examples{
data(PD)
clusterMethods(clena_result)
nClusters(clena_result)
clusters(clena_result, "kmeans")
clusters(clena_result, "hierarchical")
mat(clena_result)
summary(clena_result)
}

