% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clPage.R
\name{clPage}
\alias{clPage}
\title{Enrichment score or Similarity based on PAGE for clusters}
\usage{
clPage(genecl_obj, method, nCluster, sampleLabel = NULL, ncore = 1,
  similarity = FALSE, verbose = FALSE)
}
\arguments{
\item{genecl_obj}{a genecl or cogena object}

\item{method}{as clMethods in genecl function}

\item{nCluster}{as nClust in cogena function}

\item{sampleLabel}{sameple Label. Do make the label of interest located after
the control label in the order of factor. See details.}

\item{ncore}{the number of cores used}

\item{verbose}{verbose}
}
\description{
Enrichment score (-log2(FDR)) or Similarity based on Parametric Analysis of 
Gene Set Enrichment (PAGE) for clusters.
}
\details{
The CMap rank data are used internally (top ranked genes are Up-regulated).
}
\examples{

#annotaion
annoGMT <- "c2.cp.kegg.v5.0.symbols.gmt.xz"
annofile <- system.file("extdata", annoGMT, package="cogena")

data(Psoriasis)
clMethods <- c("hierarchical","kmeans","diana","fanny","som","model","sota","pam","clara","agnes")
genecl_result <- coExp(DEexprs, nClust=2:3, clMethods=c("hierarchical","kmeans"), 
    metric="correlation", method="complete", ncore=2, verbose=TRUE)
    
cogena_res <- clPage(genecl_result, "kmeans", "3", sampleLabel=sampleLabel)


}

