% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clusters}
\alias{clusters}
\title{get the cluster of certain clustering method.}
\usage{
clusters(object, method, ...)
}
\arguments{
\item{object}{a clena object}

\item{method}{as clMethods in clena function}
}
\description{
get the cluster of certain clustering method.
}

