% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cogena_methods.R
\docType{methods}
\name{clusterMethods}
\alias{clusterMethods}
\alias{clusters}
\alias{mat}
\alias{nClusters}
\alias{summary,cogena-method}
\title{Basic methods for a cogena object.}
\usage{
clusterMethods(object)

nClusters(object)

clusters(object, method)

mat(object)

\S4method{summary}{cogena}(object, digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}
}
\value{
clusterMethods: a character vector.

nClusters: a numeric vector.

clusters: a list or hclust depends on the method

mat: a matrix

summary: a summary of a cogena object.
}
\description{
clusterMethods: get the methods of clustering used.

nClusters: get the number of clusters from a cogena object.

clusters: get the cluster of a certain clustering method.

mat: get the original data from a cogena object.

summary: a summary of a cogena object.
}
\examples{
data(PD)
clusterMethods(cogena_result)
nClusters(cogena_result)
clusters(cogena_result, "kmeans")
clusters(cogena_result, "hierarchical")
mat(cogena_result)
summary(cogena_result)
}

