% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PD.R
\docType{data}
\name{cogena_result}
\alias{cogena_result}
\title{am example result of cogena}
\format{a cogena object.}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE20163}
}
\description{
am example result of cogena
}
\keyword{datasets}

