% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cogena.R
\docType{data}
\name{DEexprs}
\alias{DEexprs}
\title{gene expression of DEG}
\format{matrix with 1243 DEGs (row) and 17 samples (column).}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE20163}
}
\description{
gene expression of DEG
}
\keyword{datasets}

