% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enrichment}
\alias{enrichment}
\title{get the enrichment table from a clena object.}
\usage{
enrichment(object, method, nClusters, ...)
}
\arguments{
\item{object}{a clena object}

\item{method}{as clMethods in clena function}

\item{nClusters}{as nClust in clena function.}
}
\description{
get the enrichment table from a clena object.
}

