% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneDrug.R
\docType{methods}
\name{geneDrug}
\alias{geneDrug}
\alias{geneDrug,character-method}
\title{The drug and gene interaction}
\usage{
geneDrug(genenames)

\S4method{geneDrug}{character}(genenames)
}
\arguments{
\item{genenames}{the gene names. A vector.}
}
\value{
a data.frame showing drugs and the interactionTypes
}
\description{
Get the drug and gene interaction from DGIdb.
}
\examples{
data(Psoriasis)
annofile <- system.file("extdata", "c2.cp.kegg.v5.0.symbols.gmt.xz",
package="cogena")

\dontrun{
genecl_result <- coExp(DEexprs, nClust=c(2,10), clMethods=c("hierarchical","kmeans"),
    metric="correlation", method="complete", ncore=2, verbose=TRUE)

clen_res <- clEnrich(genecl_result, annofile=annofile, sampleLabel=sampleLabel)

geneC3 <- geneInCluster(clen_res, "hierarchical", "10", "3")

dgi <- drugGene(geneC3)
}
}

