% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geneInCluster.R
\docType{methods}
\name{geneInCluster}
\alias{geneInCluster}
\alias{geneInCluster,cogena-method}
\alias{geneInCluster,cogena_methods}
\title{Get gene names in a certain cluster.}
\usage{
geneInCluster(object, method, nClusters, ith)

\S4method{geneInCluster}{cogena}(object, method = clusterMethods(object),
  nClusters = nClusters(object), ith)
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}

\item{ith}{the ith cluster.}
}
\value{
a character vector containing the gene names.
}
\description{
Get gene names in a certain cluster. This is helpful if user want to get the
detail of a cluster.
}
\examples{
data(PD)
#summay this cogena object
summary(cogena_result)

#geneInCluster
geneInCluster(cogena_result, "kmeans", "3", "2")

#Up or Down genes with setting nClusters as "2".
geneInCluster(cogena_result, "som", "2", "1")
}
\seealso{
\code{\link{cogena}}
}

