% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{genesets}
\alias{genesets}
\title{gene sets.}
\format{GMT}
\source{
\url{http://www.broadinstitute.org/gsea/downloads.jsp}
}
\usage{
data(genesets)
}
\description{
genesets datasets containing the SYMBOL genes and gene sets, such as pathway, GO.
other gmt files or user-defined gene-sets file could also be used.
They are as follows:
}
\details{
\itemize{
  \item c2.cgp.v4.0.symbols.gmt. chemical and genetic perturbations, gene symbols
  \item c2.cp.kegg.v4.0.symbols.gmt. KEGG gene sets, gene symbols
  \item c2.cp.reactome.v4.0.symbols.gmt. Reactome gene sets, gene symbols
  \item c2.cp.v4.0.symbols.gmt. all canonical pathways, gene symbols
  \item c5.bp.v4.0.symbols.gmt. GO biological processes, gene symbols
  \item c6.all.v4.0.symbols.gmt. all oncogenic signatures gene sets, gene symbols
}
}
\keyword{datasets}

