% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gmt2list}
\alias{gmt2list}
\title{read gmt file as a list}
\usage{
gmt2list(anno)
}
\arguments{
\item{anno}{a gmt file. Examples are from MSigDB Collections.
The default is c2.cp.v4.0.symbols.gmt. Others like c2.cgp.v4.0.symbols.gmt,
c2.cp.kegg.v4.0.symbols.gmt, c2.cp.reactome.v4.0.symbols.gmt,
c5.bp.v4.0.symbols.gmt, c6.all.v4.0.symbols.gmt. Or a gmt format file.}
}
\description{
read Gene MatriX (gmt) file and output a list with the the first column as
the names of items in the list.
}
\examples{
anno <- "c2.cp.kegg.v4.0.symbols.gmt"
anno <- system.file("data", anno, package="clena")
gmt2list(anno)
}

