% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmtlist2file.R
\name{gmtlist2file}
\alias{gmtlist2file}
\title{write gmt list into gmt file}
\usage{
gmtlist2file(gmtlist, filename)
}
\arguments{
\item{gmtlist}{a list containing gmt}

\item{filename}{output filename}
}
\description{
write gmt list into gmt file
}
\examples{
anno <- "c2.cp.kegg.v5.0.symbols.gmt.xz"
annofile <- system.file("extdata", anno, package="cogena")
gl <- gmt2list(annofile)
gmtfile <- gmtlist2file(gl, filename="")

}
\seealso{
gmt2list
}
