% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{heatmap.PEI,clena-method}
\alias{heatmap.PEI,clena-method}
\title{heatmap.PEI}
\usage{
\S4method{heatmap.PEI}{clena}(object, method = clusterMethods(object),
  nClusters = nClusters(object), CutoffNumGeneset = 25,
  orderMethod = "max")
}
\description{
heatmap.PEI
}

