% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmapEnrich.R
\docType{methods}
\name{heatmapEnrich}
\alias{heatmapEnrich}
\alias{heatmapEnrich,cogena}
\alias{heatmapEnrich,matrix-method}
\title{heatmap of the gene set enrichment_score matrix directly (for CMap
gene set only)}
\usage{
heatmapEnrich(enrichment_score, orderMethod = "max", CutoffNumGeneset = 20,
  CutoffPVal = 0.05, low = "grey", high = "red", na.value = "white",
  title = "cogena", printGS = TRUE)

\S4method{heatmapEnrich}{matrix}(enrichment_score, orderMethod = "max",
  CutoffNumGeneset = 20, CutoffPVal = 0.05, low = "grey", high = "red",
  na.value = "white", title = "cogena", printGS = TRUE)
}
\arguments{
\item{enrichment_score}{a returned value from \code{\link{enrichment}}
function}

\item{orderMethod}{the order method, default is max, other options are
"mean", "all", "I", "II" or a number meaning the ith cluster.}

\item{CutoffNumGeneset}{the cut-off of the number of gene sets in the
return table. The default is 20.}

\item{CutoffPVal}{the cut-off of p-value. The default is 0.05.}

\item{low}{colour for low end of gradient.}

\item{high}{colour for high end of gradient.}

\item{na.value}{Colour to use for missing values.}

\item{title}{a character. Default is "cogena"}

\item{printGS}{print the enriched gene set names or not. Default is TRUE.}
}
\value{
a gene set enrichment heatmap
}
\description{
heatmap of the gene set enrichment_score matrix directly for the CMap
dataset only. This function is desgined for the cogena result from CMap
only so as to collapse the multi-isntance drugs in CMap!
After obtaining the ennrichemt of clusters in the gene
sets via \code{\link{enrichment}}, the heatmapEnrich will show it
as a heatmap.
}
\details{
This function aims to heatmap the enrichment_score directly. This is
helpful on condition that there are so many enriched gene sets and you
can filter the enrichment_score based on a criteria, like just one cluster.

orderMethod:
\itemize{
\item max. ordered by the max value in clusters beside all
\item mean. ordered by the mean value in clusters beside all
\item All. ordered by all genes
\item I. ordered by the I cluster in two clusters (Up or Down-regulated)
\item II. ordered by the II cluster in two clusters (Up or Down-regulated)
\item a number. like 2, "3".
}
}
\examples{
\dontrun{
enrichment_score <- enrichment(cogena_result, "kmeans", "3")
heatmapEnrich(enrichment_score, "kmeans", "3")
}
}

