% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{heatmapPEI}
\alias{heatmapPEI}
\title{heatmap of the gene set enrichment.}
\usage{
heatmapPEI(object, method, nClusters, CutoffNumGeneset = 20,
  CutoffPVal = 0.05, orderMethod = "max", roundvalue = TRUE,
  low = "green", high = "red", na.value = "white", ...)
}
\arguments{
\item{object}{a clena object}

\item{method}{as clMethods in clena function}

\item{nClusters}{as nClust in clena function.}

\item{CutoffNumGeneset}{the cut-off of the number of gene sets in the return table}

\item{CutoffPVal}{the cut-off of p-value. The default is 0.05.}

\item{orderMethod}{the order method, default is max, other options are "mean" and "all"}

\item{roundvalue}{whether or not round the data. such as round(1.54, 1)=1.5}

\item{low}{colour for low end of gradient.}

\item{high}{colour for high end of gradient.}

\item{na.value}{Colour to use for missing values.}
}
\description{
heatmap of the gene set enrichment index. After obtaining the ennrichemt of
clusters in the gene sets, the heatmapPEI will show it as a heatmap with an
order.
}
\details{
orderMethod:
\itemize{
\item max. ordered by the max value in clusters beside all
\item mean. ordered by the mean value in clusters beside all
\item all. ordered by all genes}
}
\examples{
#summay this clena object
summary(clena_result)

#heatmapPEI
heatmapPEI(clena_result, "kmeans", "2", orderMethod="mean")
heatmapPEI(clena_result, "kmeans", "3", CutoffNumGeneset=20,
          low = "#132B43", high = "#56B1F7", na.value = "grey50")
}
\seealso{
\code{\link{clena}} and \code{\link{heatmapCluster}}
}

