% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{heatmapPEI2}
\alias{heatmapPEI2}
\title{heatmap of the gene set enrichment matrix directly}
\usage{
heatmapPEI2(object, enrichment, ...)
}
\arguments{
\item{object}{a clena object}
}
\description{
heatmap of the gene set enrichment matrix directly. After obtaining the
ennrichemt of clusters in the gene sets via \code{\link{enrichment}}, the
heatmapPEI2 will show it as a heatmap.
}
\details{
This function aims to heatmap the enrichment score directly. This is helpful
on condition that there are so many enriched gene sets and you can filter the
enrichment score based on a criteria, like just one cluster.
}
\examples{
##
}

