% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hubGene.R
\docType{methods}
\name{hubGene}
\alias{hubGene}
\alias{hubGene,character-method}
\alias{hubGene,cluster_methods}
\title{Hub genes in a cluster}
\usage{
hubGene(geneC, score_threshold = 0)

\S4method{hubGene}{character}(geneC, score_threshold = 0)
}
\arguments{
\item{geneC}{a gene names vector}

\item{score_threshold}{a threshold for the combined scores of the interactions.}
}
\value{
the hub genes
}
\description{
Show hub genes in a cluster.
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v5.0.symbols.gmt",
package="cogena")

\dontrun{
genecl_result <- coExp(DEexprs, nClust=2:3, clMethods=c("hierarchical","kmeans"),
    metric="correlation", method="complete", ncore=2, verbose=TRUE)

clen_res <- clEnrich(genecl_result, annofile=annofile, sampleLabel=sampleLabel)
geneC3 <- geneInCluster(clen_res, "hierarchical", "3", "3")
res <- hubGene(geneC3)
}
}

