% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hubgeneInCluster.R
\docType{methods}
\name{hubgeneInCluster}
\alias{hubgeneInCluster}
\alias{hubgeneInCluster,cogena-method}
\title{Show hub gene names in certain cluster.}
\usage{
hubgeneInCluster(object, method, nCluster, ith)

\S4method{hubgeneInCluster}{cogena}(object, method = clusterMethods(object),
  nCluster = nClusters(object), ith)
}
\arguments{
\item{object}{a genecl or cogena object}

\item{method}{as clMethods in genecl function}

\item{nCluster}{as nClust in cogena function.}

\item{ith}{the ith cluster.}
}
\value{
a character vector.
}
\description{
Show hub gene names in certain cluster.
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v5.0.symbols.gmt",
package="cogena")
genecl_result <- coExp(DEexprs, nClust=2:3, clMethods=c("hierarchical","kmeans"),
    metric="correlation", method="complete", ncore=2, verbose=TRUE)

clen_res <- clEnrich(genecl_result, annofile=annofile, sampleLabel=sampleLabel)

#hubgeneInCluster
hubgeneInCluster(clen_res, "kmeans", "3", "2")
}
\seealso{
\code{\link{clEnrich}} and \code{\link{geneInCluster}}
}

