% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_methods.R
\docType{methods}
\name{mat}
\alias{mat}
\alias{mat,cogena-method}
\alias{mat,cogena_methods}
\alias{mat,genecl-method}
\alias{mat,genecl_methods}
\title{mat: get the original data from a genecl object.}
\usage{
mat(object)

\S4method{mat}{genecl}(object)

\S4method{mat}{cogena}(object)
}
\arguments{
\item{object}{a genecl or cogena object}
}
\value{
mat: a matrix
}
\description{
mat: get the original data from a genecl object.
}
\examples{
\dontrun{
mat(genecl_result)
}
}

