% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_methods.R
\docType{methods}
\name{nClusters}
\alias{nClusters}
\alias{nClusters,genecl-method}
\alias{nClusters,genecl_methods}
\alias{nClusters,cogena-method}
\alias{nClusters,cogena_methods}
\title{nClusters: get the number of clusters from a genecl object.}
\usage{
nClusters(object)

\S4method{nClusters}{genecl}(object)

\S4method{nClusters}{cogena}(object)
}
\arguments{
\item{object}{a genecl or cogena object}
}
\value{
nClusters: a numeric vector.
}
\description{
nClusters: get the number of clusters from a genecl object.
}
\examples{
\dontrun{
nClusters(genecl_result)
}
}
