% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netInCluster.R
\docType{methods}
\name{netInCluster}
\alias{netInCluster}
\alias{netInCluster,cogena-method}
\alias{netInCluster,cogena_methods}
\title{Show network in a certain cluster.}
\usage{
netInCluster(object, gn, cutoff = 0.7, ...)

\S4method{netInCluster}{cogena}(object, gn, cutoff = 0.7, ...)
}
\arguments{
\item{object}{a cogena object}

\item{gn}{A charactor vector from geneInCluster}

\item{cutoff}{similarity cut-off. Default is 0.7}

\item{...}{other parameter to igraph::plot.graph}
}
\value{
a figure.
}
\description{
It's based on igraph package.
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v4.0.symbols.gmt",
package="cogena")
cogena_result <- cogena(DEexprs, nClust=2:3,
clMethods=c("hierarchical","kmeans"), metric="correlation",
method="complete",  annofile=annofile, sampleLabel=sampleLabel,
ncore=1, verbose=TRUE)
#summay this cogena object
summary(cogena_result)

#geneInCluster
g1 <- geneInCluster(cogena_result, "kmeans", "3", "3")

netInCluster(cogena_result, g1)
}
\seealso{
\code{\link{cogena}}
}

