% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{optCluster}
\alias{optCluster}
\title{get the best clustering methods and the number of clusters}
\usage{
optCluster(object)
}
\arguments{
\item{object}{a cogena object}
}
\description{
get the best clustering methods and the number of clusters, based that the number
of gene sets which are signifigant should be maximum.
}
\examples{
summary(cogena_result)
score <- optCluster(cogena_result)
}

