% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{optClustering}
\alias{optClustering}
\title{get the best clustering methods and the number of clusters}
\usage{
optClustering(object)
}
\arguments{
\item{object}{a clena object}
}
\description{
get the best clustering methods and the number of clusters, based that the number
of gene sets which are signifigant should be maxium.
}
\examples{
##
}

