% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pClusters.R
\name{pClusters}
\alias{pClusters}
\title{paralleled clustering}
\usage{
pClusters(mat, Distmat, clMethod, nClust, method, metric, ncore, verbose, ...)
}
\value{
a list containing the cluster information
}
\description{
Internal functions.
}
\details{
Paralleled clustering with different clustering methods and different number
of clusters. These functions should not be called directly by the user.
}
\keyword{internal}

