% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPPI.R
\docType{methods}
\name{plotPPI}
\alias{plotPPI}
\alias{plotPPI,cogena-method}
\alias{plotPPI,cogena_methods}
\title{Plot the Protein-protein interaction for genes in a certain cluster.}
\usage{
plotPPI(object, method, nCluster, ith, score_threshold = 0, ...)

\S4method{plotPPI}{cogena}(object, method = clusterMethods(object),
  nCluster = nClusters(object), ith, score_threshold = 0, ...)
}
\arguments{
\item{object}{a genecl or cogena object}

\item{method}{as clMethods in genecl function}

\item{nCluster}{as nClust in cogena function.}

\item{ith}{the ith cluster.}

\item{score_threshold}{a threshold for the combined scores of the
interactions. Default is 0.}

\item{...}{parameter to plot_network from STRINGdb}
}
\value{
a figure
}
\description{
Show how genes in a certain cluster are interacted based on String database.
please note, it doesnot work if there are more than 400 genes in a cluster.
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v5.0.symbols.gmt",
package="cogena")
genecl_result <- coExp(DEexprs, nClust=2:3, clMethods=c("hierarchical","kmeans"),
    metric="correlation", method="complete", ncore=2, verbose=TRUE)

clen_res <- clEnrich(genecl_result, annofile=annofile, sampleLabel=sampleLabel)

#summay this cogena object
summary(clen_res)

#plotPPI
\dontrun{
plotPPI(clen_res, "kmeans", "3", "3")
}
}
\seealso{
\code{\link{clEnrich}}
}

