% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PPIenrichment.R
\docType{methods}
\name{PPIenrichment}
\alias{PPIenrichment}
\alias{PPIenrichment,cogena-method}
\alias{PPIenrichment,cogena_methods}
\title{enrichment for genes in a certain cluster.}
\usage{
PPIenrichment(object, method, nClusters, ith, score_threshold = 0)

\S4method{PPIenrichment}{cogena}(object, method = clusterMethods(object),
  nClusters = nClusters(object), ith, score_threshold = 0)
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}

\item{ith}{the ith cluster.}

\item{score_threshold}{a threshold for the combined scores of the
interactions. Default is 0.}
}
\value{
p-value
}
\description{
A p-value representing the enrichment in interactions of the list of
proteins (i.e. the probability to obtain such a number of interactions
by chance). It is mainly based on get_ppi_enrichment from STRINGdb.
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v4.0.symbols.gmt",
package="cogena")
cogena_result <- cogena(DEexprs, nClust=2:3,
clMethods=c("hierarchical","kmeans"), metric="correlation",
method="complete",  annofile=annofile, sampleLabel=sampleLabel,
ncore=1, verbose=TRUE)
#summay this cogena object
summary(cogena_result)

#PPIenrichment
\dontrun{
PPIenrichment(cogena_result, "kmeans", "3", "3")
}
}
\seealso{
\code{\link{cogena}}, \code{\link[STRINGdb]{get_ppi_enrichment}}
}

