% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PPIinfo.R
\docType{methods}
\name{PPIinfo}
\alias{PPIinfo}
\alias{PPIinfo,character-method}
\alias{PPIinfo,cluster_methods}
\title{Protein-protein interaction infromation for genes}
\usage{
PPIinfo(geneC, score_threshold = 0, input_directory = "", ...)

\S4method{PPIinfo}{character}(geneC, score_threshold = 0,
  input_directory = "", ...)
}
\arguments{
\item{geneC}{a gene names vector}

\item{score_threshold}{a threshold for the combined scores of the interactions.}

\item{input_directory}{a directory to store data downloaded from STRING.}

\item{...}{to get_link from STRINGdb}
}
\value{
summary of the network constructed by input genes.
}
\description{
PPI information for genes in a certain cluster. It will return a short link
 to the network page of our STRING website that shows the protein
 interactions between input genes and a p-value representing the enrichment
 in interactions of the list of vproteins (i.e. the probability to obtain
  such a number of interactions by chance). It is mainly based on
  get_link and get_summary from STRINGdb.
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v5.0.symbols.gmt",
package="cogena")

\dontrun{
genecl_result <- coExp(DEexprs, nClust=2:3, clMethods=c("hierarchical","kmeans"),
    metric="correlation", method="complete", ncore=2, verbose=TRUE)

clen_res <- clEnrich(genecl_result, annofile=annofile, sampleLabel=sampleLabel)

#PPIinfo
geneC3 <- geneInCluster(clen_res, "hierarchical", "3", "3")
PPIinfo(geneC3, input_directory="~/tmp/STRINGdb")
}
}
\seealso{
\code{\link{clEnrich}}, \code{\link[STRINGdb]{get_link}}
}

