% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogena.R
\docType{data}
\name{Psoriasis}
\alias{Psoriasis}
\alias{Psoriasis,DEexprs,sampleLabel,cogena_result}
\title{Psoriasis dataset.}
\format{two objects: DEexprs and sampleLabel.
\describe{
\item{DEexprs}{expression of DEG. There are 706 DEGs and 116 samples.}
\item{sampleLabel}{the label of sample, There are 58 control and 58 Psoriasis.}
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE13355}
}
\description{
An example dataset of Psoriasis. This dataset is used for 
illustration of the usage of \code{cogena} package. It has been normalised
the expression profling using \code{rma} method, filtered some 
non-informative genes using \code{MetaDE} package and analysed the 
differentially expressed genes using \code{limma} package with 
the cut-off adjuested p-value 0.05 and abs(logFC) >=1.
}
\keyword{datasets}

