% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PD.R
\docType{data}
\name{sampleLabel}
\alias{sampleLabel}
\title{label of samples}
\format{a vector with 17 element.}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE20163}
}
\description{
label of samples
}
\keyword{datasets}

