% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_methods.R
\docType{methods}
\name{show,cogena-method}
\alias{show,cogena-method}
\alias{show,cogena_methods}
\alias{show,genecl-method}
\title{show: show the class of cogena or genecl object}
\usage{
\S4method{show}{cogena}(object)

\S4method{show}{genecl}(object)
}
\arguments{
\item{object}{a genecl or cogena object}
}
\value{
show which instance
}
\description{
show: show the class of cogena or genecl object
}
\examples{
\dontrun{
show(genecl_result)
}
}

