% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplot.R
\name{simplot}
\alias{simplot}
\title{plotting similarity matrix}
\usage{
simplot(object, method, nCluster, orderMethod = "min",
  CutoffNumGeneset = 20, decreasing = FALSE)
}
\arguments{
\item{object}{a genecl or cogena object}

\item{method}{as clMethods in genecl function}

\item{nCluster}{as nClust in cogena function.}

\item{CutoffNumGeneset}{the cut-off of the number of gene sets in the 
return table}

\item{decreasing}{parameter for order}
}
\description{
plotting similarity matrix
}
\examples{


genecl_result <- coExp(DEexprs, nClust=2:3, clMethods=c("hierarchical","kmeans"), 
    metric="correlation", method="complete", ncore=2, verbose=TRUE)
    
annoGMT <- "c2.cp.kegg.v5.0.symbols.gmt.xz"
annofile <- system.file("extdata", annoGMT, package="cogena")

data(Psoriasis)
clMethods <- c("hierarchical","kmeans","diana","fanny","som","model","sota","pam","clara","agnes")
genecl_result <- coExp(DEexprs, nClust=10, clMethods="pam", 
    metric="correlation", method="complete", ncore=2, verbose=TRUE)   
cogena_res <- clPage(genecl_result, "pam", "10", sampleLabel=sampleLabel)
simplot(cogena_res, "pam", "10")
}

