% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stopImplicitCluster.R
\name{stopImplicitCluster}
\alias{stopImplicitCluster}
\title{stop Implicit Cluster}
\usage{
stopImplicitCluster()
}
\value{
NA
}
\description{
Internal functions.
}
\details{
Internal functions for stop cluster. This is a patch for \code{doParallel}.
Call \code{stopImplicitCluster()} after foreach/doParallel ending
}
\author{
Dan from fredhutch
}
\keyword{internal}

