% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Class_methods.R
\docType{methods}
\name{summary,genecl-method}
\alias{summary,cogena-method}
\alias{summary,cogena_methods}
\alias{summary,genecl-method}
\title{summary: a summary of a genecl object.}
\usage{
\S4method{summary}{genecl}(object)

\S4method{summary}{cogena}(object)
}
\arguments{
\item{object}{a genecl or cogena object}
}
\value{
summary: a summary of a genecl object.
}
\description{
summary: a summary of a genecl object.
}
\examples{
\dontrun{
summary(genecl_result)
}
}

