% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cogena.R
\docType{data}
\name{AllGeneSymbols}
\alias{AllGeneSymbols}
\title{All the gene symbols}
\format{a vector with 18986 gene symbols.}
\source{
\url{http://www.genenames.org/}
}
\description{
All the gene symbols
}
\keyword{datasets}

