% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGeneSymbols.R
\docType{data}
\name{AllGeneSymbols}
\alias{AllGeneSymbols}
\title{All GENE SYMBOLS}
\source{
\url{http://www.genenames.org/cgi-bin/statistics}
}
\usage{
data(AllGeneSymbols)
}
\description{
This is all the gene symbols of protein-coding genes. There are 18986 genes
totally, which are used as the backgound genes usually.
}
\details{
There are one object.
\itemize{
 \item AllGeneSymbols a vector of gene symbols.
}
}
\keyword{datasets}

