% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{clena}
\alias{clena}
\alias{clusterMethods,clena-method}
\alias{clusters,clena-method}
\alias{enrichment,clena-method}
\alias{heatmap.PEI}
\alias{mat,clena-method}
\alias{nClusters,clena-method}
\alias{summary,clena-method}
\title{cluster enrichment analysis}
\usage{
clena(obj, nClust, clMethods = "hierarchical", metric = "correlation",
  method = "complete", annotation = NULL, sampleLabel = NULL, ncore = 2,
  annotationGenesPop = NULL, verbose = FALSE, ...)

\S4method{clusterMethods}{clena}(object)

\S4method{nClusters}{clena}(object)

\S4method{clusters}{clena}(object, method = clusterMethods(object))

\S4method{mat}{clena}(object)

\S4method{enrichment}{clena}(object, method = clusterMethods(object),
  nClusters = nClusters(object))

\S4method{summary}{clena}(object, digits = max(3, getOption("digits") - 3))

heatmap.PEI(object, ...)
}
\arguments{
\item{obj}{Differentially expressed gene expression profilings. Either a
numeric matrix, a data.frame, or an ExpressionSet object. Data frames must
contain all numeric columns. In all cases, the rows are the items to be
clustered (e.g., genes), and the columns are the samples.}

\item{nClust}{A numeric vector giving the numbers of clusters to be
evaluated. e.g., 4:6 would evaluate the number of clusters ranging from 4 to 6.}

\item{clMethods}{A character vector giving the clustering methods. The
default is "hierarchical". Available options are "hierarchical", "kmeans",
"diana", "fanny", "som", "model", "sota", "pam", "clara", and "agnes",
with multiple choices allowed.}

\item{metric}{the distance measure to be used. This must be one of "euclidean",
"maximum", "manhattan", "canberra", "binary", "pearson", "abspearson",
"correlation", "abscorrelation", "spearman" or "kendall". Any unambiguous
substring can be given. In detail, please reference the parameter method in
amap::Dist. Some of the cluster methods could use only part of the metric.}

\item{method}{For hierarchical clustering (hierarchical and agnes), the agglomeration
method used. The default is "complete". Available choices are "ward", "single",
"complete", and "average". ####weighted}

\item{annotation}{logical matrix of biological annotation with row be DE gene,
column be gene sets and value be logical.}

\item{sampleLabel}{character vector with names are sample names. only used for plotting.}

\item{ncore}{Number of core used. The default is 2.}

\item{annotationGenesPop}{logical matrix of biological annotation with row be
all genes, column be gene sets and value be logical.}

\item{verbose}{verbose.}
}
\value{
a clean object
}
\description{
clena : Gene sets enrichment analysis for the cluster of gene expression profilings.
Gene sets could be pathway, GO etc.

summary : a summary of clena object.

heatmap.PEI
}
\details{
For metric, "hierarchical","kmeans","diana","fanny","pam" and "agnes" can use
all the metrics.
"clara” uses "manhattan" or "euclidean", other metric will be changed as "euclidean".
"sota" uses "correlation" or "euclidean", other metric will be changed as "euclidean".
"model" uses its own metric and "som" uses euclidean only, which is irrelative with metric.
}
\examples{
###
}

