% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cogena_methods.R
\docType{methods}
\name{clusterMethods}
\alias{clusterMethods}
\alias{clusterMethods,cogena-method}
\alias{clusterMethods,cogena_methods}
\alias{clusters}
\alias{clusters,cogena-method}
\alias{clusters,cogena_methods}
\alias{mat}
\alias{mat,cogena-method}
\alias{mat,cogena_methods}
\alias{nClusters}
\alias{nClusters,cogena-method}
\alias{nClusters,cogena_methods}
\alias{summary,cogena-method}
\title{Basic methods for a cogena object.}
\usage{
clusterMethods(object)

\S4method{clusterMethods}{cogena}(object)

nClusters(object)

\S4method{nClusters}{cogena}(object)

clusters(object, method)

\S4method{clusters}{cogena}(object, method = clusterMethods(object))

mat(object)

\S4method{mat}{cogena}(object)

\S4method{summary}{cogena}(object)
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}
}
\value{
clusterMethods: a character vector.

nClusters: a numeric vector.

clusters: a list or hclust depends on the method

mat: a matrix

summary: a summary of a cogena object.
}
\description{
clusterMethods: get the methods of clustering used.

nClusters: get the number of clusters from a cogena object.

clusters: get the cluster of a certain clustering method.

mat: get the original data from a cogena object.

summary: a summary of a cogena object.
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v4.0.symbols.gmt",
package="cogena")
cogena_result <- cogena(DEexprs, nClust=2:3,
clMethods=c("hierarchical","kmeans"), metric="correlation",
method="complete",  annofile=annofile, sampleLabel=sampleLabel,
ncore=1, verbose=TRUE)
clusterMethods(cogena_result)
\dontrun{
nClusters(cogena_result)
}
\dontrun{
clusters(cogena_result, "kmeans")
clusters(cogena_result, "hierarchical")
}
\dontrun{
mat(cogena_result)
}
\dontrun{
summary(cogena_result)
}
}

