% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corInCluster.R
\docType{methods}
\name{corInCluster}
\alias{corInCluster}
\alias{corInCluster,cogena-method}
\alias{corInCluster,cogena_methods}
\title{Correlation in the cluster of a cogena object}
\usage{
corInCluster(object, method, nClusters, ith, corMethod = "pearson",
  plotMethod = "circle", type = "upper", ...)

\S4method{corInCluster}{cogena}(object, method = clusterMethods(object),
  nClusters = nClusters(object), ith, corMethod = "pearson",
  plotMethod = "circle", type = "upper", ...)
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}

\item{ith}{the ith cluster.}

\item{corMethod}{a character string indicating which correlation coefficient
(or covariance) is to be computed. One of "pearson" (default), "kendall", or
"spearman", can be abbreviated.}

\item{plotMethod}{the visualization method of correlation matrix
to be used. Currently, it supports seven methods, named "circle" (default),
"square", "ellipse", "number", "pie", "shade" and "color". See examples in
\code{\link[corrplot]{corrplot}} for details}

\item{type}{"full" (default), "upper" or "lower", display full
matrix, lower triangular or upper triangular matrix. See examples in
\code{\link[corrplot]{corrplot}} for details}

\item{...}{other parameters to \code{\link[corrplot]{corrplot}} function.}
}
\value{
a correlation figure.
}
\description{
Correlation in the cluster of a cogena object. This is helpful if the number
of genes in cluster are small.
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v4.0.symbols.gmt",
package="cogena")
cogena_result <- cogena(DEexprs, nClust=c(2,10),
clMethods=c("hierarchical","kmeans"), metric="correlation",
method="complete",  annofile=annofile, sampleLabel=sampleLabel,
ncore=1, verbose=TRUE)
corInCluster(cogena_result, "kmeans", "10", "10")
corInCluster(cogena_result, "kmeans", "10", "10", plotMethod="square")
}
\seealso{
\code{\link{cogena}} \code{\link[corrplot]{corrplot}}
}

