% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogena.R
\docType{data}
\name{DEexprs}
\alias{DEexprs}
\title{gene expression of DEG}
\format{matrix with 706 DEGs (row) and 116 samples (column).}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE13355}
}
\description{
gene expression of DEG
}
\keyword{datasets}

