% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrichment.R
\docType{methods}
\name{enrichment}
\alias{enrichment}
\alias{enrichment,cogena-method}
\alias{enrichment,cogena_methods}
\title{get the enrichment table from a cogena object.}
\usage{
enrichment(object, method, nClusters, CutoffNumGeneset = Inf,
  CutoffPVal = 0.05, orderMethod = "max", roundvalue = TRUE)

\S4method{enrichment}{cogena}(object, method = clusterMethods(object),
  nClusters = nClusters(object), CutoffNumGeneset = Inf,
  CutoffPVal = 0.05, orderMethod = "max", roundvalue = TRUE)
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}

\item{CutoffNumGeneset}{the cut-off of the number of gene sets in the
return table}

\item{CutoffPVal}{the cut-off of p-value. The default is 0.05.}

\item{orderMethod}{the order method, default is max, other options are
"mean", "all", "I", "II" or a number meaning the ith cluster.}

\item{roundvalue}{The default is TRUE. whether or not round the data.
such as round(1.54, 1)=1.5}
}
\value{
a matrix with clusters in row and gene-sets in column.
}
\description{
get the enrichment table from a cogena object with certain clustering
methods and number of clusters.
}
\details{
orderMethod:
\itemize{
\item max. ordered by the max value in clusters beside all
\item mean. ordered by the mean value in clusters beside all
\item All. ordered by all genes
\item I. ordered by the I cluster in two clusters (Up or Down-regulated)
\item II. ordered by the II cluster in two clusters (Up or Down-regulated)
\item a number. like 2, "3".
}
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v4.0.symbols.gmt",
package="cogena")
cogena_result <- cogena(DEexprs, nClust=2:3,
clMethods=c("hierarchical","kmeans"), metric="correlation",
method="complete",  annofile=annofile, sampleLabel=sampleLabel,
ncore=1, verbose=TRUE)
enrichment.table1 <- enrichment(cogena_result, "kmeans", "3")
enrichment.table2 <- enrichment(cogena_result, "kmeans", "3",
CutoffNumGeneset=10, orderMethod="mean")
}

