% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gene2set.R
\name{gene2set}
\alias{annotationListToMatrix}
\alias{gene2set}
\title{generate relationship between genes and gene-sets}
\usage{
gene2set(anno, genenames, TermFreq = 0)

annotationListToMatrix(annotation, genenames)
}
\arguments{
\item{anno}{a gmt file. Examples are from MSigDB Collections.
A list of gene set could be find in the vignette of clena.}

\item{genenames}{a SYMBOL gene names charactic vector.}

\item{TermFreq}{a threshold for the Term Frequence. Default is zero.}
}
\description{
Generate relationship between genes (gene SYMBOL) and gene-sets, such as Pathway or GO.
}
\examples{
please ?cogena
}

