% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gene2set}
\alias{gene2set}
\title{generate relationship between genes and gene sets}
\usage{
gene2set(anno, genenames, TermFreq = 0)
}
\arguments{
\item{anno}{a string or a gmt file. The default is "c2.cp.v4.0.symbols.gmt"
from MSigDB Collections. Others like "c2.cgp.v4.0.symbols.gmt",
"c2.cp.kegg.v4.0.symbols.gmt", "c2.cp.reactome.v4.0.symbols.gmt",
"c5.bp.v4.0.symbols.gmt", "c6.all.v4.0.symbols.gmt". Or a gmt format file.}

\item{genenames}{a SYMBOL gene names charactic vector.}

\item{TermFreq}{a threshold for the Term Frequence. Default is zero.}
}
\description{
Generate relationship between genes (SYMBOL) and gene sets, such as Pathway or GO.
}
\details{
The gene sets could be gmt format. Options could be c2.cgp.v4.0.symbols.gmt,
c2.cp.reactome.v4.0.symbols.gmt,  c5.bp.v4.0.symbols.gmt,
c2.cp.kegg.v4.0.symbols.gmt  c2.cp.v4.0.symbols.gmt, c6.all.v4.0.symbols.gmt,
msigdb.v4.0.symbols.gmt. Or a user-defined gene sets gmt file.
}
\examples{
data(PD)
annoGMT <- "c2.cp.kegg.v4.0.symbols.gmt"
annoGMT <- system.file("data", annoGMT, package="clena")
anno = gene2set(anno=annoGMT, difSAM_GSE7621_DEG, TermFreq=0)
annotationGenesPop = gene2set(anno=annoGMT, rownames(GSE7621.filtered.expr), TermFreq=0)
annotationGenesPop <- annotationGenesPop[,colnames(anno)]
}

