% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gene2set}
\alias{annotationListToMatrix}
\alias{gene2set}
\title{generate relationship between genes and gene sets}
\usage{
gene2set(anno, genenames, TermFreq = 0)

annotationListToMatrix(annotation, genenames)
}
\arguments{
\item{genenames}{a SYMBOL gene names charactic vector.}

\item{TermFreq}{a threshold for the Term Frequence.}

\item{annotation}{an annotation list.}

\item{annoList}{gene set annotation gmt file. The default is c5.bp.v4.0.symbols.gmt from MSigDB Collections.}
}
\value{
an annotaion matrix.
}
\description{
Generate relationship between genes (SYMBOL) and gene sets, such as Pathway or GO.
}
\details{
The gene sets could be gmt format. Options could be c2.cgp.v4.0.symbols.gmt,
c2.cp.reactome.v4.0.symbols.gmt,  c5.bp.v4.0.symbols.gmt,
c2.cp.kegg.v4.0.symbols.gmt  c2.cp.v4.0.symbols.gmt, c6.all.v4.0.symbols.gmt,
msigdb.v4.0.symbols.gmt. Or a user-defined gene sets gmt file.
}
\examples{
###
}

