% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geneExpInCluster.R
\docType{methods}
\name{geneExpInCluster}
\alias{geneExpInCluster}
\alias{geneExpInCluster,cogena-method}
\alias{geneExpInCluster,cogena_methods}
\title{Get gene names in each clusters and the expression profiling.}
\usage{
geneExpInCluster(object, method, nClusters)

\S4method{geneExpInCluster}{cogena}(object, method = clusterMethods(object),
  nClusters = nClusters(object))
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}
}
\value{
a list containing a matrix of cluster_id with expression profiling and
label a vector of the sample labels.
}
\description{
Get gene names in each clusters and the expression profiling. This output is
helpful if user want to analyse the data for other application.
}
\examples{
\dontrun{
#summay this cogena object
summary(cogena_result)

#geneExpInCluster
geneExpInCluster(cogena_result, "kmeans", "3")
}
}
\seealso{
\code{\link{cogena}}
}

