% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{geneInCluster}
\alias{geneInCluster}
\title{Show gene names in certain cluster.}
\usage{
geneInCluster(object, method, nClusters, ith, ...)
}
\arguments{
\item{object}{a clena object}

\item{method}{as clMethods in clena function}

\item{nClusters}{as nClust in clena function.}

\item{ith}{the ith cluster.}
}
\value{
a character vector.
}
\description{
Show gene names in certain cluster.
}
\examples{
#summay this clena object
summary(GSE7621.SAM.clena.cluster)

#geneInCluster
geneInCluster(GSE7621.SAM.clena.cluster, "sota", "12", "2")
}
\seealso{
\code{\link{clena}}
}

