% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geneInCluster}
\alias{geneInCluster}
\title{Get gene names in certain cluster.}
\usage{
geneInCluster(object, method, nClusters, ith)
}
\arguments{
\item{object}{a clena object}

\item{method}{as clMethods in clena function}

\item{nClusters}{as nClust in clena function.}

\item{ith}{the ith cluster.}
}
\value{
a character vector containing the gene names.
}
\description{
Get gene names in certain cluster. This is helpful if user want to get the
detail of a cluster.
}
\examples{
#summay this clena object
summary(clena_result)

#geneInCluster
geneInCluster(clena_result, "kmeans", "3", "2")
}
\seealso{
\code{\link{clena}}
}

