% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geneInCluster.R
\name{geneInCluster}
\alias{geneInCluster}
\title{Get gene names in a certain cluster.}
\usage{
geneInCluster(object, method, nClusters, ith)
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}

\item{ith}{the ith cluster.}
}
\value{
a character vector containing the gene names.
}
\description{
Get gene names in a certain cluster. This is helpful if user want to get the
detail of a cluster.
}
\examples{
\dontrun{
#summay this cogena object
summary(cogena_result)

#geneInCluster
geneInCluster(cogena_result, "kmeans", "3", "2")
}
}
\seealso{
\code{\link{cogena}}
}

