% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heatmapCluster.R
\docType{methods}
\name{heatmapCluster}
\alias{heatmapCluster}
\alias{heatmapCluster,cogena-method}
\title{heatmap of gene expression profilings with cluster indication.}
\usage{
heatmapCluster(object, method, nClusters, sampleColor = c("darkblue", "cyan"),
  clusterColor = NULL, clusterColor2 = NULL, heatmapcol = NULL, ...)

\S4method{heatmapCluster}{cogena}(object, method = clusterMethods(object),
  nClusters = nClusters(object), sampleColor = c("darkblue", "cyan"),
  clusterColor = NULL, clusterColor2 = NULL, heatmapcol = NULL, ...)
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}

\item{sampleColor}{a color vector with the sample length. The default is c("darkblue", "cyan").}

\item{clusterColor}{a color vector with the cluster length. The default is rainbow(nClusters(object)).}

\item{clusterColor2}{a color vector with 2 elements. The default is  c("coral3", "deepskyblue1").}

\item{heatmapcol}{col for heatmap. The default is greenred(75).}

\item{...}{other parameters to heatmap.3.}
}
\description{
heatmap of gene expression profilings with cluster-based color indication.
}
\examples{
#summay this cogena object
summary(cogena_result)

#heatmapCluster
\dontrun{
heatmapCluster(cogena_result, "hierarchical", "3")
heatmapcol <- gplots::redgreen(75)
heatmapCluster(cogena_result, "hierarchical", "3", heatmapcol=heatmapcol)
}
}
\seealso{
\code{\link{cogena}}, \code{\link{heatmap.3}} and \code{\link{heatmapPEI}}
}

