% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{heatmapCluster}
\alias{heatmapCluster}
\title{heatmap of gene expression profilings with cluster indication.}
\usage{
heatmapCluster(object, method, nClusters, sampleColor = c("darkblue", "cyan"),
clusterColor = rainbow(nClusters), ...)
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}

\item{sampleColor}{a color vector with the sample length. The default is c("darkblue", "cyan").}

\item{clusterColor}{a color vector with the cluster length. The default is rainbow(nClusters(object)).}

\item{...}{other parameters to heatmap.3.}
}
\description{
heatmap of gene expression profilings with cluster-based color indication.
}
\examples{
#summay this cogena object
summary(cogena_result)

#heatmapCluster
heatmapCluster(cogena_result, "hierarchical", "3")
}
\seealso{
\code{\link{cogena}}, \code{\link{heatmap.3}} and \code{\link{heatmapPEI}}
}

