% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmapCluster.R
\docType{methods}
\name{heatmapCluster}
\alias{heatmapCluster}
\alias{heatmapCluster,cogena-method}
\title{heatmap of gene expression profilings with cluster indication.}
\usage{
heatmapCluster(object, method, nClusters, sampleColor = NULL,
  clusterColor = NULL, clusterColor2 = NULL, heatmapcol = NULL,
  maintitle = NULL, printSum = TRUE, ...)

\S4method{heatmapCluster}{cogena}(object, method = clusterMethods(object),
  nClusters = nClusters(object), sampleColor = NULL, clusterColor = NULL,
  clusterColor2 = NULL, heatmapcol = NULL, maintitle = NULL,
  printSum = TRUE, ...)
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}

\item{sampleColor}{a color vector with the sample length. The default is
from topo.colors randomly.}

\item{clusterColor}{a color vector with the cluster length.
The default is rainbow(nClusters(object)).}

\item{clusterColor2}{a color vector with 2 elements. The default is
c("coral3", "deepskyblue1").}

\item{heatmapcol}{col for heatmap. The default is greenred(75).}

\item{maintitle}{a character. like GSExxx. the output of figure will like
"kmeans 3 Clusters GSExxx" in two lines.}

\item{printSum}{print the summary of the number of genes in each cluster.
Default is TRUE.}

\item{...}{other parameters to heatmap.3.}
}
\value{
a gene expression heatmap with Cluster information figure
}
\description{
heatmap of gene expression profilings with cluster-based color indication.
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v4.0.symbols.gmt",
package="cogena")
cogena_result <- cogena(DEexprs, nClust=2:3,
clMethods=c("hierarchical","kmeans"), metric="correlation",
method="complete",  annofile=annofile, sampleLabel=sampleLabel,
ncore=1, verbose=TRUE)

#summay this cogena object
summary(cogena_result)

#heatmapCluster

heatmapCluster(cogena_result, "hierarchical", "3")
heatmapcol <- gplots::redgreen(75)
heatmapCluster(cogena_result, "hierarchical", "3", heatmapcol=heatmapcol)
}
\seealso{
\code{\link{cogena}}, \code{\link{heatmap.3}} and
\code{\link{heatmapPEI}}
}

