% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmapEnrich.R
\docType{methods}
\name{heatmapEnrich}
\alias{heatmapEnrich}
\alias{heatmapEnrich,cogena}
\alias{heatmapEnrich,cogena-method}
\title{heatmap designed for for CMap gene set only}
\usage{
heatmapEnrich(object, method = clusterMethods(object),
  nCluster = nClusters(object), orderMethod = "max",
  CutoffNumGeneset = 20, CutoffPVal = 0.05, low = "grey", high = "red",
  na.value = "white", maintitle = "cogena", printGS = TRUE, add2 = TRUE)

\S4method{heatmapEnrich}{cogena}(object, method = clusterMethods(object),
  nCluster = nClusters(object), orderMethod = "max",
  CutoffNumGeneset = 20, CutoffPVal = 0.05, low = "grey", high = "red",
  na.value = "white", maintitle = "cogena", printGS = TRUE, add2 = TRUE)
}
\arguments{
\item{object}{a genecl or cogena object}

\item{method}{as clMethods in genecl function}

\item{nCluster}{as nClust in cogena function.}

\item{orderMethod}{the order method, default is max, other options are
"mean", "all", "I", "II" or a number meaning the ith cluster.}

\item{CutoffNumGeneset}{the cut-off of the number of gene sets in the
return table. The default is 20.}

\item{CutoffPVal}{the cut-off of p-value. The default is 0.05.}

\item{low}{colour for low end of gradient.}

\item{high}{colour for high end of gradient.}

\item{na.value}{Colour to use for missing values.}

\item{maintitle}{a character. Default is "cogena"}

\item{printGS}{print the enriched gene set names or not. Default is TRUE.}

\item{add2}{enrichment score for add Up and Down reuglated genes.}
}
\value{
a gene set enrichment heatmap

orderMethod:
\itemize{
\item max. ordered by the max value in clusters beside all
\item mean. ordered by the mean value in clusters beside all
\item All. ordered by all genes
\item Up. ordered by up-regulated genes (add2 should be TRUE)
\item Down. ordered by down-regulated genes (add2 should be TRUE)
}
}
\description{
heatmapEnrich is desgined for the cogena result from CMap
only so as to collapse the multi-isntance drugs in CMap!
}
\examples{
\dontrun{
heatmapEnrich(clen_res, "kmeans", "3")
}
}

