% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmapPEI.R
\docType{methods}
\name{heatmapPEI}
\alias{heatmapPEI}
\alias{heatmapPEI,cogena}
\alias{heatmapPEI,cogena-method}
\title{heatmap of the gene set enrichment from a cogena object.}
\usage{
heatmapPEI(object, method, nCluster, CutoffNumGeneset = 20,
  CutoffPVal = 0.05, orderMethod = "max", roundvalue = TRUE,
  low = "green", high = "red", na.value = "white", maintitle = NULL,
  printGS = TRUE)

\S4method{heatmapPEI}{cogena}(object, method = clusterMethods(object),
  nCluster = nClusters(object), CutoffNumGeneset = 20, CutoffPVal = 0.05,
  orderMethod = "max", roundvalue = TRUE, low = "grey", high = "red",
  na.value = "white", maintitle = NULL, printGS = TRUE)
}
\arguments{
\item{object}{a genecl or cogena object}

\item{method}{as clMethods in genecl function}

\item{nCluster}{as nClust in cogena function.}

\item{CutoffNumGeneset}{the cut-off of the number of gene sets in the
return table}

\item{CutoffPVal}{the cut-off of p-value. The default is 0.05.}

\item{orderMethod}{the order method, default is max, other options are
"mean", "all", "I", "II" or a number meaning the ith cluster.}

\item{roundvalue}{The default is TRUE. whether or not round the data.
such as round(1.54, 1)=1.5}

\item{low}{colour for low end of gradient.}

\item{high}{colour for high end of gradient.}

\item{na.value}{Colour to use for missing values.}

\item{maintitle}{a character. like GSExxx. the output of figure will like
"cogena: kmeans 3 GSExxx" in two lines. Default is NULL}

\item{printGS}{print the enriched gene set names or not. Default is TRUE.}
}
\value{
a gene set enrichment heatmap
}
\description{
heatmap of the gene set enrichment index. After obtaining the ennrichemt of
clusters in the gene sets, the heatmapPEI will show it as a heatmap with
order.
}
\details{
orderMethod:
\itemize{
\item max. ordered by the max value in clusters beside all
\item mean. ordered by the mean value in clusters beside all
\item All. ordered by all genes
\item I. ordered by the I cluster in only two clusters
(Up or Down-regulated)
\item II. ordered by the II cluster in only two clusters
(Up or Down-regulated)
}
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v5.0.symbols.gmt",
package="cogena")

\dontrun{
genecl_result <- coExp(DEexprs, nClust=2:3, clMethods=c("hierarchical","kmeans"),
    metric="correlation", method="complete", ncore=2, verbose=TRUE)

clen_res <- clEnrich(genecl_result, annofile=annofile, sampleLabel=sampleLabel)

#summay this cogena object
summary(clen_res)

#heatmapPEI
heatmapPEI(clen_res, "kmeans", "2", orderMethod="mean")
heatmapPEI(clen_res, "kmeans", "3", CutoffNumGeneset=20,
    low = "#132B43", high = "#56B1F7", na.value = "grey50")
}
}
\seealso{
\code{\link{clEnrich}} and \code{\link{heatmapCluster}}
}

