% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{heatmapPEI}
\alias{heatmapPEI}
\title{heatmap of the gene set enrichment index.}
\usage{
heatmapPEI(object, method, nClusters, CutoffNumGeneset=20, CutoffPVal=0.05,
orderMethod="max", low="green", high="red", na.value="white")
}
\arguments{
\item{object}{a clena object}

\item{low}{colour for low end of gradient.}

\item{high}{colour for high end of gradient.}

\item{na.value}{Colour to use for missing values.}
}
\description{
heatmap of the gene set enrichment index. After obtaining the ennrichemt of
clusters in the gene sets, the heatmapPEI will show it as a heatmap with an
order.
}
\details{
orderMethod:
\itemize{
\item max. ordered by the max value in clusters beside all
\item mean. ordered by the mean value in clusters beside all
\item all. ordered by all genes}
}
\examples{
#summay this clena object
summary(GSE7621.SAM.clena.cluster)

#heatmapPEI
heatmapPEI(GSE7621.SAM.clena.cluster, "sota", "12", orderMethod="mean")
heatmapPEI(GSE7621.SAM.clena.cluster, "diana", "20", CutoffNumGeneset=20,
          low = "#132B43", high = "#56B1F7", na.value = "grey50")
}
\seealso{
\code{\link{clena}} and \code{\link{heatmapCluster}}
}

