% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heatmapPEI2.R
\docType{methods}
\name{heatmapPEI2}
\alias{heatmapPEI2}
\alias{heatmapPEI2,cogena}
\alias{heatmapPEI2,cogena-method}
\title{heatmap of the gene set enrichment_score matrix directly}
\usage{
heatmapPEI2(object, enrichment_score, method, nClusters, whichCluster,
  CutoffNumGeneset = 60, low = "grey", high = "red", na.value = "white",
  title = NULL)

\S4method{heatmapPEI2}{cogena}(object, enrichment_score, method, nClusters,
  whichCluster, CutoffNumGeneset = 60, low = "grey", high = "red",
  na.value = "white", title = NULL)
}
\arguments{
\item{object}{a cogena object}

\item{enrichment_score}{a returned value from \code{\link{enrichment}} function}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}

\item{whichCluster}{which cluster should be based to filter. The format is
"Cluster number # number of genes in clsuters", like "1#22". This can be
obtained by heatmapCluster}

\item{CutoffNumGeneset}{the cut-off of the number of gene sets in the
return table}

\item{low}{colour for low end of gradient.}

\item{high}{colour for high end of gradient.}

\item{na.value}{Colour to use for missing values.}

\item{title}{a character. like GSExxx. the output of figure will like "cogena:
kmeans 3 GSExxx" in two lines. Default is NULL}
}
\description{
heatmap of the gene set enrichment_score matrix directly. After obtaining the
ennrichemt of clusters in the gene sets via \code{\link{enrichment}}, the
heatmapPEI2 will show it as a heatmap.
}
\details{
This function aims to heatmap the enrichment_score directly. This is helpful
on condition that there are so many enriched gene sets and you can filter the
enrichment_score based on a criteria, like just one cluster.
}
\examples{
\dontrun{
data(PD)
summary(cogena_result)
enrichment.table <- enrichment(cogena_result, "kmeans", "3")

heatmapPEI2(cogena_result, enrichment.table, "kmeans", "3", "1#nn")
}
}

