% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{hubgeneInCluster}
\alias{hubgeneInCluster}
\title{Show hub gene names in certain cluster.}
\usage{
hubgeneInCluster(object, method, nClusters, ith, ...)
}
\arguments{
\item{object}{a clena object}

\item{method}{as clMethods in clena function}

\item{nClusters}{as nClust in clena function.}

\item{ith}{the ith cluster.}
}
\value{
a character vector.
}
\description{
Show hub gene names in certain cluster.
}
\examples{
#summay this clena object
summary(GSE7621.SAM.clena.cluster)

#hubgeneInCluster
hubgeneInCluster(GSE7621.SAM.clena.cluster, "sota", "12", "2")
}
\seealso{
\code{\link{clena}} and \code{\link{geneInCluster}}
}

