% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hubgeneInCluster.R
\docType{methods}
\name{hubgeneInCluster}
\alias{hubgeneInCluster}
\alias{hubgeneInCluster,cogena-method}
\title{Show hub gene names in certain cluster.}
\usage{
hubgeneInCluster(object, method, nClusters, ith)

\S4method{hubgeneInCluster}{cogena}(object, method = clusterMethods(object),
  nClusters = nClusters(object), ith)
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}

\item{ith}{the ith cluster.}
}
\value{
a character vector.
}
\description{
Show hub gene names in certain cluster.
}
\examples{
#summay this cogena object
summary(cogena_result)

#hubgeneInCluster
hubgeneInCluster(cogena_result, "kmeans", "3", "2")
}
\seealso{
\code{\link{cogena}} and \code{\link{geneInCluster}}
}

