% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optCluster.R
\docType{methods}
\name{optCluster}
\alias{optCluster}
\alias{optCluster,cogena}
\alias{optCluster,cogena-method}
\title{get the best clustering methods and the number of clusters}
\usage{
optCluster(object)

\S4method{optCluster}{cogena}(object)
}
\arguments{
\item{object}{a cogena object}
}
\description{
get the best clustering methods and the number of clusters, based that the number
of gene sets which are signifigant should be maximum.
}
\examples{
summary(cogena_result)
\dontrun{
score <- optCluster(cogena_result)
}
}

