% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optCluster.R
\docType{methods}
\name{optCluster}
\alias{optCluster}
\alias{optCluster,cogena}
\alias{optCluster,cogena-method}
\title{get the best clustering methods and the number of clusters}
\usage{
optCluster(object, based = NULL)

\S4method{optCluster}{cogena}(object, based = NULL)
}
\arguments{
\item{object}{a cogena object}

\item{based}{counting method. Default is to count all the clusters and I,
II, All. Other options are "All", "I", "II".}
}
\description{
get the best clustering methods and the number of clusters, based that the number
of gene sets which are signifigant should be maximum.
}
\examples{
data(PD)
summary(cogena_result)
\dontrun{
score <- optCluster(cogena_result)
}
}

