% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optCluster.R
\docType{methods}
\name{optCluster}
\alias{optCluster}
\alias{optCluster,cogena}
\alias{optCluster,cogena-method}
\title{get the best clustering methods and the number of clusters}
\usage{
optCluster(object, based = "inTotal", ncores = object@ncore,
  CutoffPVal = 0.05)

\S4method{optCluster}{cogena}(object, based = "inTotal",
  ncores = object@ncore, CutoffPVal = 0.05)
}
\arguments{
\item{object}{a cogena object}

\item{based}{counting method. Default is "inTotal" to count all the clusters
and I, II, All. Other options are "All", "I", "II".}

\item{ncores}{cores used for caculating optCluster. Default is same as ncores
used during cogena function, but it will be the same as number of cores
machine has if ncores parameter is exceed it.}

\item{CutoffPVal}{the cut-off of p-value. Default is 0.05.}
}
\description{
get the best clustering methods and the number of clusters, based that the number
of gene sets which are signifigant should be maximum.
}
\examples{
data(PD)
summary(cogena_result)
\dontrun{
score <- optCluster(cogena_result)
score <- optCluster(cogena_result, based="All")
}
}

