% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optCluster.R
\docType{methods}
\name{optCluster}
\alias{optCluster}
\alias{optCluster,cogena}
\alias{optCluster,cogena-method}
\title{get the best clustering methods and the number of clusters}
\usage{
optCluster(object, based = "inTotal", ncores = object@ncore,
  CutoffPVal = 0.05)

\S4method{optCluster}{cogena}(object, based = "inTotal",
  ncores = object@ncore, CutoffPVal = 0.05)
}
\arguments{
\item{object}{a genecl or cogena object}

\item{based}{counting method. Default is "inTotal" to count all enriched
gene sets from all the clusters and I, II, All. Other options are "All",
 "I", "II". Here "I" or "II" means all the up-regualted or down-regualted
 genes (Do check the output of function heatmapCluster to suggest which is
 up-regulated (down-regulated)!), while "All" means all the DEGs used.}

\item{ncores}{cores used for caculating optCluster. Default is same as ncores
used during cogena function, but it will be the same as number of cores
machine has if ncores parameter is exceed it.}

\item{CutoffPVal}{the cut-off of p-value. Default is 0.05.}
}
\value{
a score matrix
}
\description{
get the best clustering methods and the number of clusters, based that
the number of gene sets which are signifigant should be maximum.
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v5.0.symbols.gmt",
package="cogena")
genecl_result <- coExp(DEexprs, nClust=2:3, clMethods=c("hierarchical","kmeans"),
    metric="correlation", method="complete", ncore=2, verbose=TRUE)

clen_res <- clEnrich(genecl_result, annofile=annofile, sampleLabel=sampleLabel)

summary(clen_res)

score <- optCluster(clen_res)
score <- optCluster(clen_res, based="All")
}

