% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{PD}
\alias{PD}
\title{Parkinson's Disease dataset.}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE7621}
}
\usage{
data(PD)
}
\description{
Parkinson's Disease datasets contains the DEG analysed by SAM, the filtered
gene expression profilings and sample lablels. This dataset is used for
illustration of usage clena package.
}
\details{
\itemize{
 \item difSAM_GSE7621_DEG. DEG of this GSE7621, analysed by SAM
 \item GSE7621.filtered.expr. the filtered gene expression profilings
 \item sampleLabel.GSE7621. sample label in GSE7621
 \item GSE7621.SAM.DEG.expr. gene expression of DEG in GSE7621
}
}
\keyword{datasets}

