% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{PD}
\alias{PD}
\title{Parkinson's Disease dataset.}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE20163}
}
\usage{
data(PD)
}
\description{
an example dataset of Parkinson's Disease. This dataset is used for
illustration of usage clena package. Please see the vignette for the
detailed the preprocessing steps.
}
\details{
There are four objects.
\itemize{
 \item DEexprs. expression of DEG. There are 1243 DEGs and 17 samples.
 \item BGgenes. the 6122 background gene names.
 \item sampleLabel. the label of sample, There are 9 control and 8 PD.
 \item clena_result. an example of clena result.
}
}
\keyword{datasets}

