% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PD.R
\docType{data}
\name{PD}
\alias{PD}
\alias{PD,DEexprs,sampleLabel,cogena_result}
\title{Parkinson's Disease dataset.}
\format{three objects: DEexprs, sampleLabel and cogena_result.
\describe{
\item{DEexprs}{expression of DEG. There are 1243 DEGs and 17 samples.}
\item{sampleLabel}{the label of sample, There are 9 control and 8 PD.}
\item{cogena_result}{an example of cogena result.}
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE20163}
}
\description{
an example dataset of Parkinson's Disease. This dataset is used for
illustration of the usage of \code{cogena} package. It has been normalised
the expression profling using \code{rma} method, filtered some
non-informative genes using \code{MetaDE} package and analysed the
differentially expressed genes using \code{limma} package with
the p-value 0.05.
}
\keyword{datasets}

