% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PEI}
\alias{PEI}
\title{Significance of Gene sets enrichment.}
\usage{
PEI(genenames, annotation, annotationGenesPop)
}
\arguments{
\item{genenames}{a vector of gene names.}

\item{annotation}{data.frame with the gene (like all the differentially
expressed genes) in row, gene set in column.}

\item{annotationGenesPop}{data.frame with the gene in row, gene set in column.
Here genes are genes in population with filering the non-nformative genes better.}
}
\value{
a vector with P-values.
}
\description{
Caculating the significance of Gene sets enrichment based on the hypergeometric test.
This function is mainly used internally. But it can be also used directly, which
can be used for the combined clusters based on user's analysis.
}
\details{
Here the genes in annotation can be a varity of types. like all the DEG, up-regualted genes
or genes in a cluster. the gene names should be consistent with the genes in the gene sets.
}
\examples{
annoGMT <- "c2.cp.kegg.v4.0.symbols.gmt"
annofile <- system.file("data", annoGMT, package="cogena")
# the DEG gene-sets matrix
anno <- gene2set(annofile, rownames(DEexprs))
# the background gene gene-sets matrix
annotationGenesPop <- gene2set(annofile, BGgenes)
annotationGenesPop <- annotationGenesPop[,colnames(anno)]

PEI(rownames(DEexprs), anno, annotationGenesPop)
}

