% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPPI.R
\docType{methods}
\name{plotPPI}
\alias{plotPPI}
\alias{plotPPI,cogena-method}
\alias{plotPPI,cogena_methods}
\title{Plot the Protein-protein interaction for genes in a certain cluster.}
\usage{
plotPPI(object, method, nClusters, ith, score_threshold = 0, ...)

\S4method{plotPPI}{cogena}(object, method = clusterMethods(object),
  nClusters = nClusters(object), ith, score_threshold = 0, ...)
}
\arguments{
\item{object}{a cogena object}

\item{method}{as clMethods in cogena function}

\item{nClusters}{as nClust in cogena function.}

\item{ith}{the ith cluster.}

\item{score_threshold}{a threshold for the combined scores of the
interactions. Default is 0.}

\item{...}{parameter to plot_network from STRINGdb}
}
\value{
a figure
}
\description{
Show how genes in a certain cluster are interacted based on String database.
}
\examples{
data(PD)
annofile <- system.file("extdata", "c2.cp.kegg.v4.0.symbols.gmt",
package="cogena")
cogena_result <- cogena(DEexprs, nClust=2:3,
clMethods=c("hierarchical","kmeans"), metric="correlation",
method="complete",  annofile=annofile, sampleLabel=sampleLabel,
ncore=1, verbose=TRUE)
#summay this cogena object
summary(cogena_result)

#plotPPI
plotPPI(cogena_result, "kmeans", "3", "3")
}
\seealso{
\code{\link{cogena}}
}

