% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogena.R
\docType{data}
\name{sampleLabel}
\alias{sampleLabel}
\title{label of samples}
\format{a vector with 116 element.}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE13355}
}
\description{
label of samples
}
\keyword{datasets}

