% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vClusters.R
\name{vClusters}
\alias{vClusters}
\title{paralleled clustering}
\usage{
vClusters(mat, Distmat, clMethod, nClust, method, metric, annotation, ncore,
  annotationGenesPop, verbose, ...)
}
\description{
Internal cogena functions.
}
\details{
Paralleled clustering with different clustering methods and different number
of clusters. These functions should not be called directly by the user.
}
\keyword{internal}

