% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vClusters}
\alias{vClusters}
\title{paralleled clustering}
\usage{
vClusters(mat, Distmat, clMethod, nClust, method, metric, annotation, ncore,
  annotationGenesPop, verbose, ...)
}
\description{
Internal clena functions.
}
\details{
These functions are not to be called directly by the user. Paralleled
clustering with different clustering methods and different number of clusters.
}

